+++ klause/src/commonMain/kotlin/com/eignex/klause/backtrack/selector/Impact.kt
@@ -26,6 +26,6 @@ import kotlin.random.Random
  * with strong global propagators.
  */
 internal class Impact(private val maxProbes: Int = 32) : ValueSelector {
-    override fun values(session: PropagationSession, varRef: VarRef, rng: Random): Sequence<Int> =
+    override fun values(session: PropagationSession, varRef: VarRef, rng: Random): Sequence<Long> =
         probeAndOrder(session, varRef, rng, maxProbes, ascending = true)
 }
